+++ CHANGES.rst
@@ -6,3 +6,5 @@ Unreleased
 -   Ported from Flask to Quart.  This port is maintained by a single person so give some time for bugs to be ironed out, etc.
 -   For simplicity, versions will now be pinned to the latest version of sqlalchemy being targetted.
 -   Contributors document has been removed for now until I can have some bandwidth to write one.
+-   Removed track changes
+-   Add compatability for sqlalchemy 2.0